+++ repository/influxdb_test.go
@@ -1,4 +1,4 @@
-package persistance
+package repository
 
 import (
 	"log"
